requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.5
djangorestframework==3.14.0
executing==2.0.0
iniconfig==2.0.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
pluggy==1.3.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycodestyle==2.11.0
Pygments==2.16.1
pytest==7.4.2
pytest-django==4.5.2
pytest-testdox==3.1.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.1
tzdata==2023.3
wcwidth==0.2.8
